11.6 Update 3 - Updated host verification to check UD QPs, and skip VT-d disabled   check by default - Updated compiler options to enhance security of FastFabric - Improved MpiApps README file - Improved help for tools ethbw and ethextract* - Fixed HPL build for newer version oneAPI - Updated INSTALL dependency to include hwloc-libs and libatomic

+++ IbaTools/ethxmlgenerate/ethxmlgenerate.c
@@ -536,7 +536,7 @@ void getRecu_opt( int argc, char ** argv, const char *pOptShort,
 					}
 
 					if ( (argv_recu[ix] - bfParam + strlen(argv_recu[ix]) + 1) ==
-							ctCharParam )
+							(long unsigned int)ctCharParam )
 						break;
 
 				}
